+++ import_export_extensions/api/views/import_job.py
@@ -1,14 +1,15 @@
+import contextlib
 import typing
 
 from rest_framework import (
     decorators,
+    exceptions,
     mixins,
     permissions,
     response,
     status,
     viewsets,
 )
-from rest_framework.exceptions import ValidationError
 
 from ... import models, resources
 from .. import serializers
@@ -37,7 +38,7 @@ class ImportBase(type):
             return viewset
 
         # Correct specs of drf-spectacular if it is installed
-        try:
+        with contextlib.suppress(ImportError):
             from drf_spectacular.utils import extend_schema, extend_schema_view
 
             detail_serializer_class = viewset().get_detail_serializer_class()
@@ -61,8 +62,6 @@ class ImportBase(type):
                     },
                 ),
             )(viewset)
-        except ImportError:
-            pass
         return viewset
 
 
@@ -91,6 +90,15 @@ class ImportJobViewSet(
     queryset = models.ImportJob.objects.all()
     serializer_class = serializers.ImportJobSerializer
     resource_class: type[resources.CeleryModelResource] | None = None
+    search_fields = ("id",)
+    ordering = (
+        "id",
+    )
+    ordering_fields = (
+        "id",
+        "created",
+        "modified",
+    )
 
     def get_queryset(self):
         """Filter import jobs by resource used in viewset."""
@@ -147,7 +155,7 @@ class ImportJobViewSet(
         try:
             job.confirm_import()
         except ValueError as error:
-            raise ValidationError(
+            raise exceptions.ValidationError(
                 f"Wrong import job status: {job.import_status}",
             ) from error
 
@@ -165,7 +173,7 @@ class ImportJobViewSet(
         try:
             job.cancel_import()
         except ValueError as error:
-            raise ValidationError(
+            raise exceptions.ValidationError(
                 f"Wrong import job status: {job.import_status}",
             ) from error